+++ firmware/robot2015/src-ctrl/modules/ControllerTaskThread.cpp
@@ -105,7 +105,7 @@ void Task_Controller(void const* args) {
 
     std::vector<uint16_t> duty_cycles;
     duty_cycles.assign(5, 100);
-    for (int i = 0; i < duty_cycles.size(); ++i)
+    for (size_t i = 0; i < duty_cycles.size(); ++i)
         duty_cycles.at(i) = 100 + 206 * i;
 
     duty_cycles.at(1) = 10;
